renames `floor-sweeper-poll` to `tribute-tools`
