+++ src/applications/tribute-tools/commands/sweep.ts
