+++ src/applications/tribute-tools/handlers/endedPollsHandler.ts
