+++ src/applications/tribute-tools/handlers/endedPollsHandler.unit.test.ts
