+++ src/applications/tribute-tools/handlers/pollReactionHandler.ts
